Fix warnings in dependencies

+++ src/3rd party/crypto/openssl/src/b_print.c
@@ -386,7 +386,7 @@ _dopr(
                 if (cflags == DP_C_SHORT) {
                     short int *num;
                     num = va_arg(args, short int *);
-                    *num = currlen;
+                    *num = (short)currlen;
                 } else if (cflags == DP_C_LONG) { /* XXX */
                     long int *num;
                     num = va_arg(args, long int *);
